requirements.txt
# Unified requirements for Django, Streamlit, LangChain, FAISS, etc.
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Django==5.2
faiss-cpu==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-huggingface==0.1.2
langchain-nvidia-ai-endpoints
langchain-text-splitters==0.3.8
langsmith==0.3.31
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
yarl==1.19.0
zstandard==0.23.0
